OCMAKU-39 move code filter to AbstractCategoryFilterState
OCMAKU-39 move code filter to AbstractCategoryFilterState
